LayoutTests/storage/indexeddb/modern/opendatabase-request-event-expected.txt

This test calls open on window.indexedDB in various ways, verifying that IDBOpenDBRequest objects are returned when expected and exceptions are thrown when expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Caught exception TypeError: Type error
ALERT: Caught exception TypeError: Value -1 is outside the range [0, 9007199254740991]
ALERT: [object IDBOpenDBRequest]
ALERT: upgradeneeded: old version - 0 new version - 1
PASS successfullyParsed is true

TEST COMPLETE
